refactor: remove providers and replace cached-resources by current-user
