--- src/cached-resources/cached-resources-context.js
@@ -1,8 +1,0 @@
-import { createContext } from 'react'
-
-const CachedResourcesContext = createContext({
-    me: {},
-    systemInfo: {},
-})
-
-export { CachedResourcesContext }
