--- src/cached-resources/index.js
@@ -1,2 +1,0 @@
-export { CachedResourcesProvider } from './cached-resources-provider.js'
-export { useCurrentUser, useSystemVersion, useIsAuthorized } from './hooks.js'
