+++ src/current-user/current-user-context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+const CurrentUserContext = createContext({})
+
+export { CurrentUserContext }
