+++ src/current-user/current-user-provider.test.js
@@ -2,7 +2,7 @@ import { useDataQuery } from '@dhis2/app-runtime'
 import { shallow, mount } from 'enzyme'
 import React from 'react'
 import { expectRenderError } from '../test-utils/expect-render-error.js'
-import { CachedResourcesProvider } from './cached-resources-provider.js'
+import { CurrentUserProvider } from './current-user-provider.js'
 
 jest.mock('@dhis2/app-runtime', () => ({
     useDataQuery: jest.fn(),
@@ -12,13 +12,11 @@ afterEach(() => {
     jest.resetAllMocks()
 })
 
-describe('<AuthWall>', () => {
+describe('<CurrentUserProvider>', () => {
     it('shows a spinner when loading', () => {
         useDataQuery.mockImplementation(() => ({ loading: true }))
 
-        const wrapper = mount(
-            <CachedResourcesProvider>Child</CachedResourcesProvider>
-        )
+        const wrapper = mount(<CurrentUserProvider>Child</CurrentUserProvider>)
         const loadingIndicator = wrapper.find({
             'data-test': 'dhis2-uicore-circularloader',
         })
@@ -36,7 +34,7 @@ describe('<AuthWall>', () => {
             error,
         }))
 
-        expectRenderError(<CachedResourcesProvider {...props} />, message)
+        expectRenderError(<CurrentUserProvider {...props} />, message)
     })
 
     it('renders the children once data has been received', () => {
@@ -45,14 +43,11 @@ describe('<AuthWall>', () => {
             error: undefined,
             data: {
                 me: {},
-                systemInfo: {
-                    version: '2.37.0',
-                },
             },
         }))
 
         const wrapper = shallow(
-            <CachedResourcesProvider>Child</CachedResourcesProvider>
+            <CurrentUserProvider>Child</CurrentUserProvider>
         )
 
         expect(wrapper.text()).toEqual(expect.stringContaining('Child'))
