+++ src/current-user/use-current-user.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react'
+import { CurrentUserContext } from './current-user-context'
+
+export const useCurrentUser = () => useContext(CurrentUserContext)
